eliminate a file; ./test passes except for runtime
