+++ enclone_print/src/lib.rs
@@ -12,7 +12,6 @@ pub mod print_utils2;
 pub mod print_utils3;
 pub mod print_utils4;
 pub mod print_utils5;
-pub mod proc_cvar1;
 pub mod proc_cvar_auto;
 pub mod proc_lvar1;
 pub mod proc_lvar2;
